Make MainForm/GaugeDemo references consistent
Also make the GaugeDemo form the default startup form. Will later move the gauge from the other form to this one, and delete the other form.
